+++ packages/expo-permissions/src/ExpoPermissions.web.ts
@@ -240,6 +240,8 @@ async function getPermissionAsync(
       };
     }
     case 'location':
+    case 'locationForeground':
+    case 'locationBackground':
       {
         const maybeStatus = await getPermissionWithQueryAsync('geolocation');
         if (maybeStatus) {